Refactor posts package
