+++ apps/trpfrog.net/src/app/blog/[slug]/og-image/route.tsx
@@ -1,4 +1,3 @@
-import { fetchBlogPost } from '@trpfrog.net/posts'
 import consola from 'consola'
 import { ImageResponse } from 'next/og'
 import { ImageResponseOptions, NextRequest } from 'next/server'
