+++ packages/posts/src/core/index.ts
@@ -1,0 +1,2 @@
+export { type BlogPost, type BlogFrontMatter, BlogPostSchema } from './blogPost.ts'
+export { searchBlogPost } from './search.ts'
